Make markdown-lint compatible changes
Largely we have to change the list formatting, so I've made the
numbering compatible with the sub-list bullets.

The parser didn't like the numbering of the different reporting
examples, so respecting that idea, and what we can get out of them by
being able to reference examples numerically, I have suffixed each
list item with a [UCXXX] value for our own reporting.

Links have all been changed to reference style and can be found in
alphabetical order at the end of the document.

Ping @peterVG