Allow Jekyll to organize pages automatically
This commit uses Jekyll's mechanics to help us to conditionally
organize pages in the accepted log based on those records that are
either proposed or accepted. Front-matter has been updated per-page
as well as information in the template. The commit happens to fix
a small issue whereby CONTRIBUTING.md wasn't rendering, but there is
no reason it shouldn't. Finally, because of this change, any updates
to ADRs do not also need to be added manually to the accepted log.
New records will be picked-up by the index and displayed. Maintenance
should be easier this way.